+++ src/freight_fate/states/driving.py
@@ -190,6 +190,7 @@ class DrivingState(DrivingControlsMixin, DrivingUpdateMixin, DrivingEventMixin, 
         self._ramp_light_last_phase = ""  # "red" | "yellow" | "green", once announced
         self._ramp_terminal_done = False
         self._ramp_waiting_at_light = False
+        self._ramp_creep_prompt_said = False
         self._destination_exit_taken = False
         self._missed_destination_exit_said = False
         self._destination_exit_announced_key = ""